refactor(models): unify W-derived helpers on _SpatialModelBase
Phase 4c: lift _W_pt_sparse, _T_ww, _batch_mean_row_sum, and
_batch_mean_row_sum_MW into the shared _SpatialModelBase parent.  All
four are byte-for-byte identical between the cross-sectional and panel
bases except for which sparse-W attribute _W_pt_sparse wraps; that
choice is now controlled by a single class hook _pt_sparse_attr
('_W_sparse' for cross-section, '_W_sparse_NT' for panel).

957/957 tests pass.
