+++ bayespecon/models/_common.py
@@ -59,6 +59,12 @@ class _SpatialModelBase(ABC):
     _decision_spec_attr: str = "get_spec"
     _diagnostics_require_W: bool = True
 
+    # Name of the sparse-W attribute that :attr:`_W_pt_sparse` should
+    # wrap as a PyTensor sparse variable.  Cross-sectional models point
+    # this at ``_W_sparse`` (the n×n weight matrix); panel models
+    # override to ``_W_sparse_NT`` (the (N·T)×(N·T) Kronecker form).
+    _pt_sparse_attr: str = "_W_sparse"
+
     # ------------------------------------------------------------------
     # Abstract interface — implemented by leaf subclasses
     # ------------------------------------------------------------------
@@ -637,6 +643,118 @@ class _SpatialModelBase(ABC):
         overrides this to enforce the check.
         """
         return None
+
+    # ------------------------------------------------------------------
+    # Shared W-derived helpers
+    # ------------------------------------------------------------------
+
+    @property
+    def _W_pt_sparse(self):
+        """PyTensor sparse variable wrapping the model's sparse W.
+
+        Cached so repeated PyMC model builds reuse the same symbolic
+        sparse operator and avoid the ``O(n²)`` dense materialisation
+        that ``pt.as_tensor_variable(self._W_dense)`` performs each
+        time.  The sparse attribute being wrapped is controlled by the
+        :attr:`_pt_sparse_attr` class hook (``"_W_sparse"`` for
+        cross-section, ``"_W_sparse_NT"`` for panel).
+
+        Use with :func:`pytensor.sparse.structured_dot` (vector inputs
+        must first be reshaped to ``(n, 1)`` because the vector
+        overload's backward pass is broken in PyTensor).
+        """
+        if not hasattr(self, "_W_pt_sparse_cache") or self._W_pt_sparse_cache is None:
+            import scipy.sparse as _sp
+            from pytensor import sparse as _pts
+
+            W_attr = getattr(self, self._pt_sparse_attr)
+            self._W_pt_sparse_cache = _pts.as_sparse_variable(_sp.csc_matrix(W_attr))
+        return self._W_pt_sparse_cache
+
+    @property
+    def _T_ww(self) -> float:
+        """Trace of W'W + W², cached on first access.
+
+        Computed as ``||W||_F² + sum(W * W')`` using sparse operations,
+        which is O(nnz) rather than O(n²).
+        """
+        if not hasattr(self, "_T_ww_cache"):
+            from ..graph import sparse_trace_WtW_plus_WW
+
+            self._T_ww_cache = sparse_trace_WtW_plus_WW(self._W_sparse)
+        return self._T_ww_cache
+
+    def _batch_mean_row_sum(self, rho_draws: np.ndarray) -> np.ndarray:
+        """Compute mean row sum of ``(I - ρW)^{-1}`` for each posterior draw.
+
+        For row-standardised W this is the scalar ``1/(1 - ρ)``.
+        For non-row-standardised W the eigenvalue decomposition is used:
+        ``mean_row_sum = (1/n) * ones' V diag(1/(1-ρω)) V^{-1} ones``,
+        where the vector ``c = V^{-1} ones`` is pre-computed once.
+
+        Parameters
+        ----------
+        rho_draws : np.ndarray, shape (G,)
+            Spatial autoregressive parameter draws.
+
+        Returns
+        -------
+        np.ndarray, shape (G,)
+            Mean row sum for each draw.
+        """
+        if self._is_row_std:
+            return 1.0 / (1.0 - rho_draws)
+
+        # Eigenvalue-based computation: precompute c = V^{-1} @ ones once.
+        if not hasattr(self, "_eig_inv_ones"):
+            W_dense = self._W_dense
+            eigs, V = np.linalg.eig(W_dense)
+            self._W_eigs_full = eigs.real.astype(np.float64)
+            self._V_full = V.real.astype(np.float64)
+            self._eig_inv_ones = np.linalg.solve(
+                self._V_full, np.ones(W_dense.shape[0])
+            )
+
+        c = self._eig_inv_ones
+        eigs = self._W_eigs_full
+        V_col_sums = self._V_full.sum(axis=0)  # (n,)
+        from ..diagnostics.spatial_effects import _chunked_eig_means
+
+        return _chunked_eig_means(rho_draws, eigs, weights=V_col_sums * c)
+
+    def _batch_mean_row_sum_MW(self, rho_draws: np.ndarray) -> np.ndarray:
+        """Compute mean row sum of ``(I - ρW)^{-1} W`` for each posterior draw.
+
+        For row-standardised W this equals ``1/(1 - ρ)`` (same as
+        :meth:`_batch_mean_row_sum`) because row sums of ``M @ W`` equal
+        row sums of ``M`` when W is row-standardised.
+
+        For non-row-standardised W the eigenvalue decomposition is used:
+        ``mean_row_sum_MW = (1/n) * ones' V diag(ω/(1-ρω)) V^{-1} ones``.
+
+        Parameters
+        ----------
+        rho_draws : np.ndarray, shape (G,)
+            Spatial autoregressive parameter draws.
+
+        Returns
+        -------
+        np.ndarray, shape (G,)
+            Mean row sum of ``M @ W`` for each draw.
+        """
+        if self._is_row_std:
+            return 1.0 / (1.0 - rho_draws)
+
+        # Ensure eigenvalue decomposition is available
+        if not hasattr(self, "_eig_inv_ones"):
+            _ = self._batch_mean_row_sum(rho_draws[:1])
+
+        c = self._eig_inv_ones
+        eigs = self._W_eigs_full
+        V_col_sums = self._V_full.sum(axis=0)  # (n,)
+        from ..diagnostics.spatial_effects import _chunked_eig_means
+
+        return _chunked_eig_means(rho_draws, eigs, weights=eigs * V_col_sums * c)
 
     # ------------------------------------------------------------------
     # Internals
