+++ bayespecon/models/base.py
@@ -419,115 +419,6 @@ class SpatialModel(_SpatialModelBase):
             dtype=np.float64,
         )
 
-    @property
-    def _W_pt_sparse(self):
-        """PyTensor sparse variable wrapping :attr:`_W_sparse`.
-
-        Cached so repeated PyMC model builds reuse the same symbolic sparse
-        operator and avoid the ``O(n²)`` dense materialisation that
-        ``pt.as_tensor_variable(self._W_dense)`` performs each time.
-
-        Use with :func:`pytensor.sparse.structured_dot` (vector inputs must
-        first be reshaped to ``(n, 1)`` because the vector overload's
-        backward pass is broken in PyTensor).
-        """
-        if not hasattr(self, "_W_pt_sparse_cache") or self._W_pt_sparse_cache is None:
-            import scipy.sparse as _sp
-            from pytensor import sparse as _pts
-
-            self._W_pt_sparse_cache = _pts.as_sparse_variable(
-                _sp.csc_matrix(self._W_sparse)
-            )
-        return self._W_pt_sparse_cache
-
-    @property
-    def _T_ww(self) -> float:
-        """Trace of W'W + W², cached on first access.
-
-        Computed as ``||W||_F² + sum(W * W')`` using sparse operations,
-        which is O(nnz) rather than O(n²).
-        """
-        if not hasattr(self, "_T_ww_cache"):
-            from ..graph import sparse_trace_WtW_plus_WW
-
-            self._T_ww_cache = sparse_trace_WtW_plus_WW(self._W_sparse)
-        return self._T_ww_cache
-
-    def _batch_mean_row_sum(self, rho_draws: np.ndarray) -> np.ndarray:
-        """Compute mean row sum of (I - rho*W)^{-1} for each posterior draw.
-
-        For row-standardised W this is the scalar ``1/(1 - rho)``.
-        For non-row-standardised W the eigenvalue decomposition is used:
-        ``mean_row_sum = (1/n) * ones' V diag(1/(1-rho*omega)) V^{-1} ones``,
-        where the vector ``c = V^{-1} ones`` is pre-computed once.
-
-        Parameters
-        ----------
-        rho_draws : np.ndarray, shape (G,)
-            Spatial autoregressive parameter draws.
-
-        Returns
-        -------
-        np.ndarray, shape (G,)
-            Mean row sum for each draw.
-        """
-        if self._is_row_std:
-            return 1.0 / (1.0 - rho_draws)
-
-        # Eigenvalue-based computation: precompute c = V^{-1} @ ones once.
-        if not hasattr(self, "_eig_inv_ones"):
-            W_dense = self._W_dense
-            eigs, V = np.linalg.eig(W_dense)
-            self._W_eigs_full = eigs.real.astype(np.float64)
-            self._V_full = V.real.astype(np.float64)
-            self._eig_inv_ones = np.linalg.solve(
-                self._V_full, np.ones(W_dense.shape[0])
-            )
-
-        c = self._eig_inv_ones
-        eigs = self._W_eigs_full
-        # Closed form per draw:
-        # mean_row_sum(rho) = (1/n) * sum_i (V_col_sums_i * c_i) / (1 - rho*omega_i)
-        # Compute in chunks over draws to bound memory at O(chunk*n).
-        V_col_sums = self._V_full.sum(axis=0)  # (n,)
-        from ..diagnostics.spatial_effects import _chunked_eig_means
-
-        return _chunked_eig_means(rho_draws, eigs, weights=V_col_sums * c)
-
-    def _batch_mean_row_sum_MW(self, rho_draws: np.ndarray) -> np.ndarray:
-        """Compute mean row sum of (I - rho*W)^{-1} W for each posterior draw.
-
-        For row-standardised W this equals ``1/(1 - rho)`` (same as
-        ``_batch_mean_row_sum``) because row sums of M@W = row sums of M
-        when W is row-standardised.
-
-        For non-row-standardised W the eigenvalue decomposition is used:
-        ``mean_row_sum_MW = (1/n) * ones' V diag(omega/(1-rho*omega)) V^{-1} ones``.
-
-        Parameters
-        ----------
-        rho_draws : np.ndarray, shape (G,)
-            Spatial autoregressive parameter draws.
-
-        Returns
-        -------
-        np.ndarray, shape (G,)
-            Mean row sum of M@W for each draw.
-        """
-        if self._is_row_std:
-            return 1.0 / (1.0 - rho_draws)
-
-        # Ensure eigenvalue decomposition is available
-        if not hasattr(self, "_eig_inv_ones"):
-            _ = self._batch_mean_row_sum(rho_draws[:1])
-
-        c = self._eig_inv_ones
-        eigs = self._W_eigs_full
-        V_col_sums = self._V_full.sum(axis=0)  # (n,)
-        from ..diagnostics.spatial_effects import _chunked_eig_means
-
-        return _chunked_eig_means(rho_draws, eigs, weights=eigs * V_col_sums * c)
-
     # ------------------------------------------------------------------
     # Input parsing helpers
     # ------------------------------------------------------------------
